requirements.txt
python-binance
pandas
numpy
python-dotenv
TA-Lib
requests